requirements.txt
requests
numpy
pillow
pathos
tifffile
opencv-python
marshmallow
shapely
rtree
networkx
pandas
